[Mac] Disable the WebM format reader in configurations where the plug-in will fail to load https://bugs.webkit.org/show_bug.cgi?id=219938
Reviewed by Eric Carlson.

The user default for loading an ad-hoc code-signed WebM format reader plug-in is not
respected in customer installs of macOS, leaving no known way to load the plug-in in an
engineering build running on a customer OS variant. Fixing this is tracked by
rdar://72320419, but for now we must disable the plug-in for this configuration.

* platform/graphics/avfoundation/objc/MediaPlayerPrivateAVFoundationObjC.mm:
(WebCore::isFormatReaderAvailable):
(WebCore::MediaPlayerPrivateAVFoundationObjC::supportsTypeAndCodecs):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270890 268f45cc-cd09-0410-ab3c-d52691b4dbfc