[Debug][GStreamer] Crash in fast/mediastream/change-tracks-media-stream-being-played.html https://bugs.webkit.org/show_bug.cgi?id=219437
Patch by Víctor Manuel Jáquez Leal <vjaquez@igalia.com> on 2020-12-16
Reviewed by Philippe Normand.

The crash was a product of a race condition where a new sample, either video or audio,
appeared after the track was already removed from the stream and the GStreamer's source was
already removed. The result was a segmentation fault when the sample observer is called,
since the observer is coupled with the GStreamer's source.

This fix consists of removing the observers as soon as the track is removed from the stream.

No new tests needed since it fix one.

* platform/mediastream/gstreamer/GStreamerMediaStreamSource.cpp:
(webkitMediaStreamSrcRemoveTrackObserver): New function.
(stopObservingTracks): Use new function.
(webkitMediaStreamSrcSetupSrc): Use track's type method rather than source's.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@270906 268f45cc-cd09-0410-ab3c-d52691b4dbfc