REGRESSION (r262729): Poor contrast for specific color/background combinations on disabled input fields https://bugs.webkit.org/show_bug.cgi?id=219991 <rdar://problem/70522921>
Reviewed by Sam Weinig.

Source/WebCore:

The text color for disabled input fields is adjusted based on any
specified color and background-color rules.

r262729 modified the adjustment logic to use the contrastRatio
utility function to determine illegibility issues due to contrast,
rather than a differenceSquared computation. In both cases if the
computed value is less than a minimum constrast value, the original
text color is preserved. As a part of the change, the minimum
contrast value was set to 1.1, a value chosen in order to maintain
the appearance of fast/forms/input-disabled-color.html.

However, the constrast value 1.1 results in combination of
"color #acacac" and "background-color: #f2f2f2" producing white
text on a light gray background, which is illegible. Previously,
this color combination did not result in any change to the text color.

In order to fix, increase the minimum color constrast to 1.195,
which accounts for the failing combination. Note that this increase to
the minimum value prevents some existing text color adjustments
from being performed. This is evidenced by the need to rebaseline
fast/forms/input-disabled-color.html after increasing the constrast.
However, the new text colors are subjectively more legible, which
is the reason why we have a minimum color constrast value.

* rendering/RenderTheme.cpp:
(WebCore::RenderTheme::disabledTextColor const):

LayoutTests:

Updated an existing test to account for the failing color/background
combination and rebaselined expectations to match the new minimum
constrast ratio.

* fast/forms/input-disabled-color.html:
* platform/ios/fast/forms/input-disabled-color-expected.txt:
* platform/mac-catalina/fast/forms/input-disabled-color-expected.txt:
* platform/mac-mojave/fast/forms/input-disabled-color-expected.txt:
* platform/mac/fast/forms/input-disabled-color-expected.txt:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271021 268f45cc-cd09-0410-ab3c-d52691b4dbfc