Retain IPC::Connection when using it when making a NetworkDataTask https://bugs.webkit.org/show_bug.cgi?id=220067
Patch by Alex Christensen <achristensen@webkit.org> on 2020-12-21
Reviewed by Geoffrey Garen.

r265230 helped, but it's still crashing.
Maybe we're getting a non-null pointer to an object that is about to be destroyed on another thread.
There's no reason not to keep it alive here.
Also, don't release assert in isParentProcessAFullWebBrowser if there's no audit token so we don't
just kick this can down the road.

* NetworkProcess/cocoa/NetworkSessionCocoa.mm:
(WebKit::NetworkSessionCocoa::sessionWrapperForTask):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271040 268f45cc-cd09-0410-ab3c-d52691b4dbfc