MediaStreamTrack.muted value should not change before the corresponding event is fired https://bugs.webkit.org/show_bug.cgi?id=220084
Reviewed by Eric Carlson.

Source/WebCore:

Before the patch, the muted value was changed and a task was queued to dispatch the corresponding event.
During a period of time, the muted value was changed while the event was not fired.
To fix this, dispatch a task that will change the value and dispatch the event.
MediaStreamTrack is now storing the muted value instead of relying on its m_private muted value.
The muted value is only used for JS as the rest of the code base expects to get the actual muted value.

Test: fast/mediastream/media-stream-track-muted-event.html

* Modules/mediastream/MediaStreamTrack.cpp:
(WebCore::MediaStreamTrack::MediaStreamTrack):
(WebCore::MediaStreamTrack::mutedForBindings const):
(WebCore::MediaStreamTrack::trackMutedChanged):
* Modules/mediastream/MediaStreamTrack.h:
* Modules/mediastream/MediaStreamTrack.idl:

LayoutTests:

* fast/mediastream/media-stream-track-muted-event-expected.txt: Added.
* fast/mediastream/media-stream-track-muted-event.html: Added.
* platform/ios/mediastream/audio-muted-in-background-tab.html: Updated to cope with async muted change.
* platform/ios/mediastream/video-muted-in-background-tab.html: Dittod


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271060 268f45cc-cd09-0410-ab3c-d52691b4dbfc