Add a helper method to WTF::MachSemaphore to wait with a timeout duration https://bugs.webkit.org/show_bug.cgi?id=220110
Reviewed by Sam Weinig.

Adds a helper method on the cross-process WTF::MachSemaphore class (to be utilized in webkit.org/b/219586).

* wtf/cocoa/MachSemaphore.cpp:
(WTF::MachSemaphore::waitFor):

This wraps a call to `semaphore_timedwait`, and converts the given time (in WTF::Seconds) into a
`mach_timespec_t`, which is an unsigned representing the number of seconds, along with another integer
representing the number of nanoseconds.

* wtf/cocoa/MachSemaphore.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271101 268f45cc-cd09-0410-ab3c-d52691b4dbfc