Copy Localizable.strings file into the correct directory https://bugs.webkit.org/show_bug.cgi?id=220029
Reviewed by Tim Horton.

After the changes in r269865, r269873, r269915, r269932, r269937 the localizable.strings
was not showing up in the right place in the production build to allow it to be correctly
localized. This skips the step of putting it in DerivedSources and just puts the
concatenated file in the final correct directory so we do not have to relay on Xcode build
doing the correct thing for us.

* WebCore.xcodeproj/project.pbxproj:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271188 268f45cc-cd09-0410-ab3c-d52691b4dbfc