REGRESSION (r261157): Crash in WKSelectPopover when running as iPhone app on iPad https://bugs.webkit.org/show_bug.cgi?id=220065 <rdar://problem/71932792>
Reviewed by Darin Adler.

Source/WebKit:

r261157 changed how WebKit determined the user interface idiom for
WKWebViews that were created in daemons. Since daemons do not create
UIApplications, WebKit cannot use UIDevice in these instances. Instead,
the user interface idiom determination logic was updated to include
a check for device hardware, using MobileGestalt, in cases where no
UIApplication was created.

Since WebKit only determines the user interface idiom once (then
storing the obtained value), the added determination logic breaks
down for iPhone apps on iPad. Consider the following sequence of
events, eventually leading to a crash when interacting with a
<select> element:

1. A WKWebView is created prior to UIApplication initialization. This
   can be achieved by creating one in another class's "load" method.

2. Since the app is physically running on an iPad, WebKit determines
   the user interface idiom to be "iPad" and saves this information.

3. Once the app actually launches, UIApplication is initialized.
   Since this is an iPhone app on iPad, the actual user interface
   idiom is "iPhone". However, WebKit does not know this, since
   it uses the saved user interface idiom from (2).

4. When tapping a <select> element, WebKit checks its saved idiom,
   an attempts to present a UIPopoverController (the standard behavior
   under the iPad idiom).

5. However, since the actual idiom is iPhone, UIKit throws an
   NSInvalidArgumentException, since UIPopoverController
   should not be used when running under the iPhone idiom.

A simple fix for the crash would be to call into UIKit to check the
idiom each time it is required, rather than relying on the saved bit.
However, this does not address the more general issue, which is that
WebKit's saved idiom and the actual idiom can be out of sync.

Consequently, the approach taken is to update the saved idiom when
it changes. This should only occur when a UIApplication is initialized.
Hence, we listen for UIApplicationDidFinishLaunchingNotification,
and if the actual idiom is different from our saved idiom, we update the
saved idiom and also notify the WebProcess.

* Shared/UserInterfaceIdiom.h:
* Shared/UserInterfaceIdiom.mm:
(WebKit::updateCurrentUserInterfaceIdiom):

This method checks the actual idiom and updates the saved idiom if
they differ. Returns true only if the idiom was updated.

* UIProcess/Cocoa/WebProcessPoolCocoa.mm:
(WebKit::WebProcessPool::registerNotificationObservers):

Listen for UIApplicationDidFinishLaunchingNotification if the idiom
can change (UIApplication is uninitialized). Then, if the actual
and saved idiom differ, notify all WebProcesses of the change.

(WebKit::WebProcessPool::unregisterNotificationObservers):
* UIProcess/WebProcessPool.h:
* WebProcess/WebProcess.h:
* WebProcess/WebProcess.messages.in:
* WebProcess/cocoa/WebProcessCocoa.mm:
(WebKit::WebProcess::userInterfaceIdiomDidChange):

Update the saved idiom using the information from the UIProcess.

Tools:

Added an API test to exercise the previously crashing, and now fixed,
codepath. Note that the test will trivially pass on iPhones, since the
crash only occurs on iPads. See below for an explanation of how the
test functions on iPads.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/ios/UserInterfaceIdiomUpdate.mm: Added.
(TestWebKitAPI::TEST):

Since TestWebKitAPI is not an application, [UIApplication sharedApplication]
is nil when the test creates a WKWebView. Consequently, when the WKWebView
is created, WebKit uses the device hardware information to determine the user
interface idiom (iPad). Now, before the test interacts with the WKWebView,
the user interface idiom is changed to an iPhone and the UIApplication is
initialized. Note that the UIApplicationDidFinishLaunchingNotification
is manually posted, since there is no UIApplicationDelegate.

Without the changes in this patch, the test will crash, since the user
interface idiom maintained by WebKit would not be updated, and would
remain iPad rather than iPhone. Hence, when focusing a select element, the test
would attempt to present a UIPopoverController. However, since the idiom
is actually iPhone, UIKit will throw an NSInvalidArgumentException.

After the changes in this patch, the test does crash, as the change to the
user interface idiom after UIApplication creation is recognized by WebKit,
and the attempt to present a UIPopoverController under the iPhone idiom
does not occur.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271190 268f45cc-cd09-0410-ab3c-d52691b4dbfc