[iOS] REGRESSION(r265088): "pointerdown" doesn't fire using a trackpad after double-tapping https://bugs.webkit.org/show_bug.cgi?id=220072
Reviewed by Antoine Quint.

r265088 made it such that pointer events are not created for mouse events if there is an
existing entry for any touch event in `m_activePointerIdsToCapturingData`. Unfortunately,
entries only appear to be removed from `m_activePointerIdsToCapturingData` from three places:
 - when the single tap gesture recognizer resets (`-[WKContentView _singleTapDidReset:]`)
 - if a potential tap cannot be committed (`-[WKContentView _commitPotentialTapFailed]`)
 - after a synthetic click (`-[WKContentView _didCompleteSyntheticClick]`)
AFAICT (and seeing how there's a gesture recognizer for double-tap, long press, etc.), this
does not include other situations like the second tap of a double-tap. In order to fix this:
 - eagerly `touchWithIdentifierWasRemoved` when dispatching `"pointerup"` for a touch event
 - (just in case) don't prevent the pointer event for mouse events if the existing touch
   event has been cancelled or is not currently pressed

* page/PointerCaptureController.cpp:
(WebCore::PointerCaptureController::dispatchEventForTouchAtIndex):
(WebCore::PointerCaptureController::pointerEventForMouseEvent):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271216 268f45cc-cd09-0410-ab3c-d52691b4dbfc