Make sure that if NetworkProcess clears DOMCache, it also clears service worker registrations https://bugs.webkit.org/show_bug.cgi?id=220408 <rdar://problem/72360003>
Reviewed by Alex Christensen.

It is expected that clearing DOM cache mandates clearing service worker registrations
as service workers expect cache entries added at install time to stay.
It is hard for clients to enforce this so it is best to handle it in Network Process.
When DOMCache is cleared, service worker registrations will also be cleared.
The reverse is not guaranteed.
Covered by API test.

* NetworkProcess/NetworkProcess.cpp:
(WebKit::NetworkProcess::deleteWebsiteData):
(WebKit::NetworkProcess::deleteWebsiteDataForOrigins):
(WebKit::NetworkProcess::deleteAndRestrictWebsiteDataForRegistrableDomains):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@271291 268f45cc-cd09-0410-ab3c-d52691b4dbfc